requirements.txt
asttokens==3.0.1
decorator==5.2.1
et_xmlfile==2.0.0
executing==2.2.1
greenlet==3.3.1
ipython==9.10.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
matplotlib-inline==0.2.1
numpy==2.4.2
openpyxl==3.1.5
pandas==3.0.1
parso==0.8.6
pexpect==4.9.0
playwright==1.58.0
prompt_toolkit==3.0.52
ptyprocess==0.7.0
pure_eval==0.2.3
pyee==13.0.1
Pygments==2.19.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
six==1.17.0
stack-data==0.6.3
traitlets==5.14.3
typing_extensions==4.15.0
wcwidth==0.6.0
